requirements.txt
bcrypt==3.1.0
PyJWT==1.5.0
enum34
futures
mock
six
flask
scipy
scikit-learn
requests
waitress
tensorflow
grpcio
werkzeug
